% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracyfunction.R
\name{accuracyfunction}
\alias{accuracyfunction}
\title{Function to evaluate simulation accuracy.}
\usage{
accuracyfunction(x)
}
\description{
Subordinate function to hybridpowercomp used in dplyr::do(.) when evaluating simulation accuracy.
}
