% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridpower.R
\name{hybridpower}
\alias{hybridpower}
\title{Power evaluation for NewHybrids analysis of simulated datasets}
\usage{
hybridpower(dir, filetag = "", Threshold = NULL, samplesize = 200)
}
\arguments{
\item{dir}{path directory which holds the output from different runs through New Hybrids (e.g. 3 simulations with 3 replicate runs each through NH) note that this directory should only hold the output folders.}

\item{filetag}{A name tag which will be added to the outputs}

\item{Threshold}{A threshold which will be added to the plots showing the assignment success for different levels of probability of a given class estimated by NewHybrids. Default is (NULL) so if nothing is specified it will not add this to the output plots (success ~ threshold by class)}

\item{samplesize}{The number of simulated individuals per NewHybrid class (default = 200). The assumption is there is equal numbers for each class. If not there must be a 0 or NA added hold the data structure}
}
\description{
\code{hybridpower} Evaluates the accuracy with which NewHybrids assigns individuals of known hybrid class to the correct hybrid class in simulated datasets at varying levels of stringency (PofZ). The code will write graphical and numerical results to the directory provided by the user.
}

