% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridPowerComp_temp.R
\name{hybridPowerComp}
\alias{hybridPowerComp}
\title{Assignment power comparison among different SNP subsets using NewHybrids simulated datasets3.}
\usage{
hybridPowerComp(dir, save_output = TRUE, return_workspace = FALSE,
  Thresholds = c(0.5, 0.6, 0.7, 0.8, 0.9), samplesize = NULL, CT = 0.1,
  CTI = 0.5)
}
\arguments{
\item{dir}{File path to the directory in which the NewHybrids results (in individual folders as returned by parallelNH_XX) are located.}

\item{save_output}{A logical indicating whether the plots and plotting data should be saved to the hard drive. The default is TRUE}

\item{return_workspace}{A logical indicating whether the plots and plotting data should be returned to the workspace as a list object. The default is FALSE}

\item{Thresholds}{A vector of critical posterior probability thresholds by which to facet plots. The default is c(0.5,0.6,0.7,0.8,0.9)}

\item{samplesize}{The number of individuals per NewHybrids class. By (default: NULL) this data will be extracted from the "*individuals.txt" output from parallelnewhybrids if present in the same folder as the PofZ file. This can also explicitly defined as a vector (6 values corresponding to # in P1,P2,F1,F2,BC1,BC2) or a path to a *_Individuals.txt.}

\item{CT}{The threshold posterior probability of assignment (PofZ) to F2 above which Pure Population 1 or Pure Population 2 individuals are flagged to indicate possible non-convergence. The default is 0.1.}

\item{CTI}{The proportion of individuals in either Pure Population 1 OR Pure Population 2 allowed to exceed the F2 assignment threshold (PofZCutOff). The default is 0.5.}
}
\description{
Evaluates the accuracy with which NewHybrids assigns individuals of known hybrid class to the correct hybrid class in simulated datasets at varying levels of stringency (PofZ). The code will write graphical and numerical results to the directory provided by the user.
}
