% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridPowerComp3.R
\name{hybridPowerComp3}
\alias{hybridPowerComp3}
\title{Assignment power comparison among different SNP subsets using NewHybrids simulated datasets3.}
\usage{
hybridPowerComp3(dir, filetag = "", Thresholds = c(0.5, 0.6, 0.7, 0.8, 0.9),
  addThresh = FALSE, samplesize = NULL, CT = 0.1, CTI = 0.5)
}
\arguments{
\item{dir}{File path to the directory in which the NewHybrids results (in individual folders as returned by parallelNH_XX) are located.}

\item{filetag}{An optional character vector to be applied as the name of the outputs.}

\item{Thresholds}{A vector of thresholds which will be added to the plots showing the assignment success for different levels of probability of a given class estimated by NewHybrids. Default is (NULL) so if nothing is specified it will not add this to the output plots (success ~ threshold by class).}

\item{samplesize}{The number of individuals per NewHybrids class. By (default: NULL) this data will be extracted from the "*individuals.txt" output from parallelnewhybrids if present in the same folder as the PofZ file. This can also explicitly defined as a vector (6 values corresponding to # in P1,P2,F1,F2,BC1,BC2) or a path to a *_Individuals.txt.}

\item{CT}{The threshold posterior probability of assignment (PofZ) to F2 above which Pure Population 1 or Pure Population 2 individuals are flagged to indicate possible non-convergence. The default is 0.1.}

\item{CTI}{The proportion of individuals in either Pure Population 1 OR Pure Population 2 allowed to exceed the F2 assignment threshold (PofZCutOff). The default is 0.5.}
}
\description{
Evaluates the accuracy with which NewHybrids assigns individuals of known hybrid class to the correct hybrid class in simulated datasets at varying levels of stringency (PofZ). The code will write graphical and numerical results to the directory provided by the user.
}

