% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_class.R
\name{n class}
\alias{n class}
\alias{n_class}
\title{Evalute the number of simulated individuals per NewHybrid class#'}
\usage{
n_class(x)
}
\arguments{
\item{x}{vector of sample IDs provided by the *_Individuals.txt output from \code{nh_analysis_data_generatoR}. This must be a vector and not a df.}
}
\description{
\code{n_class} will provide a count of each class according to vector of sample names specified in the *_individuals.txt output from  \code{nh_analysis_data_generatoR}
}
