% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_accuracy_checkR.R
\docType{data}
\name{nh_accuracy_checkR}
\alias{NHresults}
\alias{nh_accuracy_checkR}
\title{NewHybrids accuracy checker}
\format{An object of class \code{data.frame} with 324 rows and 7 columns.}
\usage{
NHresults
}
\arguments{
\item{NHResults}{a dataframe that is the PofZ object}

\item{print.results}{a logical query for whether the results should be printed to the screen in addition to exported as an object. Default is TRUE}

\item{all.hyb}{a logical query if the proportion of all indivuals known to be hybrids were assigned to a hybrid category regardless if the category was the correct one. Default is FALSE}
}
\description{
\code{nh_accuracy_checkR} calcultes the accuracy with which NewHybrids assigns (simulated) individuals to their known genotype frequency category. That is, what proportion of all individuals that are known to belong to a given category were classified as belonging to that category by NH. The proportion correct will be returned as a dataframe with the name "NH.accuracy", and can also be printed to the screen if desired
}
\keyword{datasets}

