% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_build_Example_Results.R
\name{nh_build_example_results}
\alias{nh_build_example_results}
\title{Creates a folder containing example NewHybrids analyses which can be evaluated using hybridpowercomp}
\usage{
nh_build_example_results(dir = NULL, remove_example = FALSE)
}
\arguments{
\item{dir}{File path to the directory in write the NewHybrids results (in individual folders as returned by parallelNH_XX). The default is NULL, in which case the folder will be written to the working directory}

\item{remove_example}{A logical indicating if the example data should be removed from the users hard drive. Default = FALSE. NOTE: If a dir was previously specified, it must be specified again}
}
\description{
Creates a folder containing example NewHybrids analyses which can be evaluated using hybridpowercomp
}
