% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_multiplotR.R
\docType{data}
\name{nh_multiplotR}
\alias{NHResults}
\alias{nh_multiplotR}
\title{NewHybrids Multi-Plot}
\format{An object of class \code{character} of length 1.}
\usage{
NHResults
}
\arguments{
\item{NHResults}{the directory in which the NewHybrids results (in individual folders as returned by parallelNH_XX) are located.}
}
\description{
\code{nh_plotR} plots the cumulative probabilities of assignment for each individual for multiple NewHybrids results
}
\keyword{datasets}

