% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_zcore copy.R
\name{nh_zcore}
\alias{nh_zcore}
\title{NewHybrids genotype category designator}
\usage{
nh_zcore(GetstheZdir, multiapplyZvec = NULL, applyuniqueZvec = NULL)
}
\arguments{
\item{GetstheZdir}{A file path to the folder in which the NewHybrids formatted files to be appended reside}

\item{multiapplyZvec}{IF a single file of Zvecs is to be applied to each file, USE multiapplyZvec. Zvecs will added in batch fashion to every NewHybrids format file within a working folder - multiapplyZvec must be specified as a file path + file name (e.g. "~/HoldenUniversity/CoolEthansComputer/HairDolls/Zvexxx.csv")}

\item{applyuniqueZvec}{IF each NewHybrids file is to be given a UNIQUE vector of Zvecs use applyuniqueZvec ---- NOTE <- to apply UNIQUE Zvecs, the Zvec files must all be placed in a single folder separate from the NewHybrids files AND they MUST follow the file name convention "NHFileName_Zvec.csv" where NHFileName is the same name as the file to which the Zvec is to be applied - consequently, the number of NewHybrids files = number of Zvec files applyuniqueZvec must be spedifed as a file path to the folder in which the Zvec files live (e.g. "~/HarrisonUniversity/DEANGordonPritchard/LettersOfRecommendation/")}
}
\description{
\code{nh_zcore} is a function that will insert a column of known genotype categories <- Zvec into a NewHybrids formatted dataset for information on the known genotype categories, refer to the NewHybrids documentation (e.g. Pure1 = z0, Pure2 = z1, F1 = z2 etc. [numbering follows that of NewHybrids]) fucntion will match a vector of categories (Zvec) to individuals, then output a NH format file, with a name indicating that the category vector has been added. The user must provide the Zvec(s) to the function in the form of a two column .csv file. The vector of genotype categories NEED NOT be the same length as the NewHybrids file - thus, time can be saved. NOTE Zvec file must have column names, and the first column must be named "Individual" and contain the names of individuals to which categories must be given that match those in the first column of the NewHybrids dataset. NOTE both multiapllyZvec and applyuniqueZvec cannot be used at the same time.
}

