% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preCheckR.R
\name{preCheckR}
\alias{preCheckR}
\title{Annotate Classes}
\usage{
preCheckR(PreDir)
}
\arguments{
\item{PreDir}{the directory in which the NewHybrids results (in individual folders as returned by parallelNH_XX) are located}
}
\description{
\code{preCheckR} detects cases where the NewHybrids MCMC chains fail to converge, and warns the user that the identified analyses should be repeated.
}

