% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type1function.R
\name{type1function}
\alias{type1function}
\title{Function to evaluate type I error.}
\usage{
type1function(x)
}
\arguments{
\item{x}{input dataframe created during hybridpowercomp}
}
\description{
Subordinate function to hybridpowercomp used in dplyr::do(.) when evaluating simulation type I error.
}

