% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type2function.R
\name{type2function}
\alias{type2function}
\title{Function to evaluate simulation type II error.}
\usage{
type2function(x)
}
\description{
Subordinate function to hybridpowercomp used in dplyr::do(.) when evaluating simulation type II error.
}
