% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_multiplotR.R
\name{nh_multiplotR}
\alias{nh_multiplotR}
\title{NewHybrids Multi-Plot}
\usage{
nh_multiplotR(NHResults)
}
\arguments{
\item{NHResults}{the directory in which the NewHybrids results (in individual folders as returned by parallelNH_XX) are located.}
}
\description{
\code{nh_plotR} plots the cumulative probabilities of assignment for each individual for multiple NewHybrids results
}
