% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_plotR.R
\name{nh_plotR}
\alias{nh_plotR}
\title{NewHybrids Plot}
\usage{
nh_plotR(nh_output)
}
\arguments{
\item{nh.output}{the PofZ file from a NewHybrids analysis}
}
\description{
\code{nh_plotR} plots the cumulative probabilities of assignment for each individual
}

