% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_plotR.R
\name{nh_plotR}
\alias{nh_plotR}
\title{NewHybrids Plot}
\usage{
nh_plotR(NHResults)
}
\arguments{
\item{NHResults}{A file path to the NewHybrids (PofZ file) result to be plotted}
}
\description{
\code{nh_plotR} plots the cumulative probabilities of assignment for each individual
}

