% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_subsetR.R
\name{nh_subset}
\alias{nh_subset}
\title{Subset NewHybrids format datasets}
\usage{
nh_subset(NHData, loci = NULL)
}
\arguments{
\item{NHData}{The file path to a NewHybrids formatted dataset to be subsetted}

\item{loci}{A vector of Loci names to be retained from the dataset}
}
\description{
Function allows the user to subset NewHybrids format datasets to explore the efficacy of variously sized panels
}

