% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NH_Zcore.R
\name{NH_Zcore}
\alias{NH_Zcore}
\title{Annotate Classes}
\usage{
NH_Zcore(GetstheZdir, multiapplyZvec = NULL, applyuniqueZvec = NULL)
}
\arguments{
\item{GetstheZdir}{A file path to the folder in which the files to be annotated reside}

\item{multiapplyZvec}{IF a single file of Zvecs is to be applied to each file, USE multiapplyZvec multiapplyZvec must be specified as a file path + file name (e.g. "~/HoldenUniversity/CoolEthansComputer/HairDolls/Zvexxx.csv")}

\item{applyuniqueZvec}{IF each NewHybrids file is to be given a UNIQUE vector of Zvecs use applyuniqueZvec ---- NOTE <- to apply UNIQUE Zvecs, the Zvec files must all be placed in a single folder separate from the NewHybrids files AND they MUST follow the file name convention "NHFileName_Zvec.csv" where NHFileName is the same name as the file to which the Zvec is to be applied - consequently, the number of NewHybrids files = number of Zvec files applyuniqueZvec must be spedifed as a file path to the folder in which the Zvec files live (e.g. "~/HarrisonUniversity/DEANGordonPritchard/LettersOfRecommendation/")}
}
\description{
\code{NH_Zcore} adds the Zed to ....
}

